do more in the actions build separate out steps so that logs are distinct

+++ .github/projects/morphia/build.sh
@@ -2,12 +2,12 @@
 
 set -e
 
-export MVN="$(pwd)/mvnw"
-
-[ -d core/src/main/java ] && $MVN install -DskipTests
-rm -rf core/src/main/java
-cp ../.github/projects/morphia/core-pom.xml core/pom.xml
-rewrite.sh
-
-echo -ne "\033]30;Compiling\007"
-$MVN -e -f core/pom.xml test-compile
+cd git_repo
+if [ -d core/src/main/java ]
+then
+  ./mvnw install -DskipTests
+  rm -rf core/src/main/java
+  cp ../.github/projects/morphia/core-pom.xml core/pom.xml
+else
+  ./mvnw -e -f core/pom.xml test-compile
+fi
